Add manage wallet context
